requirements.txt
amqp==5.1.1
arrow==0.11.0
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-appconf==1.0.4
django-celery-beat==2.2.1
django-celery-email==3.0.0
django-celery-results==2.2.0
django-ckeditor==6.2.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-countries==7.2.1
django-environ==0.8.1
django-extensions==3.1.0
django-htmlmin==0.11.0
django-js-asset==2.0.0
django-mail-queue==3.2.5
django-object-actions==3.0.1
django-pagedown==2.2.1
django-phonenumber-field==6.0.0
django-python3-ldap==0.11.3
django-rest-passwordreset==1.3.0
django-timezone-field==4.2.3
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-extensions==0.6.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
et-xmlfile==1.0.1
flower==1.0.0
form==0.0.1
gunicorn==20.1.0
html5lib==1.1
humanize==4.1.0
ics==0.7
idna==3.3
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.1.2
kombu==5.2.4
ldap3==2.8.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
openpyxl==3.0.5
packaging==21.3
parse==1.19.0
pdfkit==0.6.1
phonenumbers==8.12.51
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
Pygments==2.7.3
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pyuca==1.2
PyYAML==5.3.1
redis==4.3.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
TatSu==4.4.0
toml==0.10.2
tornado==6.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
XlsxWriter==1.3.7
